+++ changelog.rst
@@ -4,6 +4,11 @@ Compysition changelog
 Version
 1.3.0
 
+- Added a greenlet local variable which is intended to set a single path of actor to actor communication.
+	- Added PendingDeprecation to Actor.send_event, Actor.send_error, Actor._clear_all, Actor._send, and Actor._loop_send
+	- Having a single point of communication can open the door to improved actor to actor communication via reducing (and in some cases removing) the need for serialization of events.
+	- Added unittests that support the new and old methods of sending events between Actors.
+	- Refactored Actor methods to improve code reusability and readability.
 - Added new event types to better support requests with 'application/x-www-form-urlencoded' mime-types.
     - Adds ability to respond to requests in 'application/x-www-form-urlencoded' format.
     - Includes special event types for JSON/XML data passed via 'application/x-www-form-urlencoded'.  However the use of these special types is NOT the default behavior. They optional via an HTTPServer parameter to preserve functionality.